requirements.txt
tweepy == 4.14.0
notion-client == 2.2.1
python-dotenv == 1.0.1